プロンプト変更。バッテリー対応

+++ sg_laucher.pyw
@@ -16,7 +16,7 @@ def load_commands(filename):
 commands_details = load_commands('sg_launcher_config.json')
 
 # コマンドを非同期で実行する関数
-def run_command(command, parameters, window):
+def run_command(command, parameters, window,wait_sw=True):
     global stop_thread
     try:
         env = os.environ.copy()
@@ -24,22 +24,26 @@ def run_command(command, parameters, window):
         command_list = command.split() + parameters.split()
         process = subprocess.Popen(command_list, stdout=subprocess.PIPE, stderr=subprocess.PIPE,  text=True, bufsize=1, universal_newlines=True,env=env)
 
-        while True:
-            if stop_thread:
-                process.terminate()
-                window.write_event_value('-THREAD_STOPPED-', 'スレッドが停止しました')
-                stop_thread = False
-                break
-            output = process.stdout.readline()
-            if output:
-                print(output)
-                window.write_event_value('-COMMAND_OUTPUT-', output.strip())
-            elif process.poll() is not None:
-                break
-            error = process.stderr.read()
-            if error:
-                print(error)
-                window.write_event_value('-COMMAND_ERROR-', error)
+        if wait_sw:
+            while True:
+                if stop_thread:
+                    process.terminate()
+                    window.write_event_value('-THREAD_STOPPED-', 'スレッドが停止しました')
+                    stop_thread = False
+                    break
+                output = process.stdout.readline()
+                if output:
+                    print(output,end='' )
+                    window.write_event_value('-COMMAND_OUTPUT-', output.strip())
+                
+                elif process.poll() is not None:
+                    error = process.stderr.read()
+                    if error:
+                        print("エラー内容",error)
+                        window.write_event_value('-COMMAND_ERROR-', error)
+                        raise Exception("呼出先でエラーです。")
+                    elif process.poll() is not None:
+                        break
 
 
     except Exception as e:
@@ -61,15 +65,12 @@ initial_command_description = initial_command['description']
 command_column = sg.Column([
     [sg.Text('コマンド:', size=(12, 1)), sg.Combo(list(commands_details.keys()), default_value=list(commands_details.keys())[0], key='-COMMAND-', readonly=True,enable_events=True)]
 ])
-
 command_body_column = sg.Column([
     [sg.Text('コマンド実体:', size=(12, 1)), sg.InputText(initial_command_body,key='-COMMAND_BODY-', readonly=True, size=(25, 1))]
 ])
-
 params_column = sg.Column([
-    [sg.Text('パラメーター:', size=(12, 1)), sg.InputText(initial_command_params,key='-PARAMS-', readonly=True, size=(25, 1))]
+    [sg.Text('パラメーター:', size=(12, 1)), sg.Multiline(initial_command_params,key='-PARAMS-', size=(50, 3))]
 ])
-
 
 layout = [
     [command_column],
@@ -98,6 +99,9 @@ while True:
         window['-COMMAND_BODY-'].update(selected_command['command'])
         window['-PARAMS-'].update(selected_command['params'])
         window['-DESCRIPTION-'].update(selected_command['description'])
+        params_allow_input = selected_command.get('params_allow_input', True)  # デフォルトはTrue
+        print(params_allow_input)
+        window['-PARAMS-'].update(disabled=not params_allow_input)        
 
     if event == '-START-':
         # 開始ボタン: 無効, 停止ボタン: 有効
@@ -105,9 +109,11 @@ while True:
         window['-STOP-'].update(disabled=False)        
         selected_command_name = values['-COMMAND-']
         selected_command = commands_details[selected_command_name]['command']
-        selected_params = commands_details[selected_command_name]['params']
+        user_params = values['-PARAMS-']  # ユーザーが入力したパラメータ
+        should_wait = commands_details[selected_command_name].get('wait', True)  # waitキーがなければTrueをデフォルト値とする
+        # 'params_allow_input' に基づいてパラメータ入力フィールドの有効/無効を切り替える
         # スレッドを開始する
-        threading.Thread(target=run_command, args=(selected_command, selected_params, window), daemon=True).start()
+        threading.Thread(target=run_command, args=(selected_command, user_params, window,should_wait), daemon=True).start()
         window['-OUTPUT-'].update(f'コマンドを投入しました{selected_command_name}....\n')
 
     elif event == '-STOP-':
